+++ src/lib.rs
@@ -2,7 +2,6 @@ pub mod asm;
 pub mod parser;
 pub mod sem;
 pub mod syntax;
-pub mod tokenizer;
 mod util;
 use sem::SemanticAnalyzer;
