+++ src/syntax.rs
@@ -1,8 +1,10 @@
 mod errors;
 mod parser;
+mod tokenizer;
 pub mod traverse;
 mod tree;
 
 pub use errors::{ParseError, ParseErrorKind};
 pub use parser::Parser;
+pub use tokenizer::*;
 pub use tree::*;
